modules/src/storage-api/requirements.txt
functions-framework==3.*
flask==2.*
google-cloud-storage==2.*
google-cloud-secretmanager==2.*